Added news module
