+++ index.js
@@ -7,9 +7,10 @@ app.use(express.json());
 
 const port = process.env.PORT || 3000;
 app.listen(port, () => console.log(`we are listening on ${port}`));
-app.use((err, req, res, next) => {
+
+routes(app);
+
+app.use((err) => {
     console.log(err);
-    res.send(err.message);
     res.status(500);
 });
-routes(app);
